feat: increment spring boot version to 3.1.0

+++ src/main/resources/db/migration/V1.003__Create_course_and_relations.sql
@@ -1,7 +1,7 @@
 create sequence course_seq start with 1 increment by 50;
-create table course (id int8 not null, name varchar(255), primary key (id));
-create table course_students (courses_id int8 not null, students_id int8 not null, primary key (courses_id, students_id));
-create table course_teachers (courses_id int8 not null, teachers_id int8 not null, primary key (courses_id, teachers_id));
+create table course (id bigint not null, name varchar(255), primary key (id));
+create table course_students (courses_id bigint not null, students_id bigint not null, primary key (courses_id, students_id));
+create table course_teachers (courses_id bigint not null, teachers_id bigint not null, primary key (courses_id, teachers_id));
 alter table if exists course_students add constraint FK532dg5ikp3dvbrbiiqefdoe6m foreign key (students_id) references student;
 alter table if exists course_students add constraint FKh6irfl8rj4jgb3xrlyxsr2bdk foreign key (courses_id) references course;
 alter table if exists course_teachers add constraint FKe3n62rwx3uc1yucgkmw6gjkm5 foreign key (teachers_id) references teacher;
